feat(admin-vue3): 重构字典管理页面
- 重新设计了字典管理页面的布局和功能
- 新增了字典组和字典项的分开管理
- 优化了树形结构和表格的展示效果
- 增加了刷新缓存功能
- 调整了相关组件的命名和结构
